functions/requirements.txt
firebase-functions>=0.1.2
firebase-admin>=6.4.0
rembg>=2.0.62
Pillow>=10.4.0
onnxruntime>=1.19.2
